IOT-33: Split nmea parser from gnss uart handling
